requirements.txt
pyopenssl
pandas
pyttsx3
pycrypto
prompt_toolkit
pyyaml